refactor: move all Rust tests from src/ to separate tests/ folder
Move all inline #[cfg(test)] modules from 15 source files to 14
separate test files under rust/tests/. This makes test presence
and implementation clearly visible from the directory structure.

- Extract normalize_url to html module as public function
- Make clean_markdown public for external test access
- All 103 tests pass, clippy clean, cargo fmt applied
